services/generalizer/requirements.txt
transformers
accelerate
torch
llama-cpp-python
loguru
fastapi
uvicorn[standard]
PyYAML
pydantic-settings